requirements.txt
uwsgi==2.0.11.1
Pygments==1.6
Django==1.7.1
requests==2.3.0
django-bootstrap3==4.11.0
django-endless-pagination
versioneer==0.13
djangorestframework==3.2.4
django-assets==0.11